+++ packages/web3-rpc-providers/CHANGELOG.md
@@ -36,3 +36,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 
 ## [Unreleased]
+ - When error is returned with code 429, throw rate limit error (#7102)